Fix UUID unittest (#20996)
UUID is now handled by the adapter therefore is added to options sent to
the server from the adapter.

Since the UUID is no longer being added by the server, the server tests
should no longer test if the server is correctly handling the UUID. This
logic is now passed to the adapter and therefore tests related to
processing non-existent UUID and UUID not currently active is the
adapters role.